refactor overhaul
