+++ src/index.ts
@@ -1,9 +1,9 @@
-import { ConsentService } from './services/consent-service';
-import { CookieConsentTranslations } from './services/translation-service';
-import { CookieConsentTheme } from './services/theme-service';
-import { CookieConsentSettings } from './services/settings-service';
-import { initConsentModal } from './init/initConsentModal';
-import { initConsentBadge } from './init/initConsentBadge';
+import { ConsentService } from '@/services/consent-service';
+import { CookieConsentTranslations } from '@/services/translation-service';
+import { CookieConsentTheme } from '@/services/theme-service';
+import { CookieConsentSettings } from '@/services/settings-service';
+import { initConsentModal } from '@/init/initConsentModal';
+import { initConsentBadge } from '@/init/initConsentBadge';
 
 type WindowCookieConsent = ConsentService;
