+++ src/themes/zoot.ts
@@ -1,4 +1,4 @@
-import { CookieConsentTheme } from '../services/theme-service';
+import { CookieConsentTheme } from '@/services/theme-service';
 
 export const themeZoot: CookieConsentTheme = {
   /* ZOOT is default default theme */
