+++ src/utils/handleBodyAnchorClicks.ts
@@ -1,12 +1,15 @@
-import { EVENT_CLICK } from '../config';
+import { EVENT_CLICK } from '@/config';
 
 interface TextAnchorClickProps {
   onAnchorClick?: (event: Event) => void;
 }
 
-export const handleAnchorClicks = <T extends TextAnchorClickProps>(bodyEl: HTMLElement, { onAnchorClick }: T): void => {
+export const handleAnchorClicks = <T extends TextAnchorClickProps, E extends HTMLElement>(
+  bodyEl: E,
+  { onAnchorClick }: T,
+): E => {
   if (onAnchorClick === undefined || typeof onAnchorClick !== 'function') {
-    return;
+    return bodyEl;
   }
 
   bodyEl.addEventListener(EVENT_CLICK, (event): void => {
@@ -19,4 +22,6 @@ export const handleAnchorClicks = <T extends TextAnchorClickProps>(bodyEl: HTMLE
       onAnchorClick(event);
     }
   });
+
+  return bodyEl;
 };
