+++ src/utils/read-consent.ts
@@ -1,4 +1,4 @@
-import { COOKIE_NAME } from '../config';
+import { COOKIE_NAME } from '@/config';
 import { ConsentRules } from './consent';
 import { getCookieByName } from './cookies';
